requirements.txt
sounddevice
librosa
numpy
scipy
lxml
openai-whisper
soundfile 